pyproject.toml
[project]
name = "speedmonitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "influxdb-client>=1.49.0",
    "python-dotenv>=1.2.1",
    "schedule>=1.2.2",
    "speedtest-cli>=2.1.3",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.3",
]
